requirements-web.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cssselect2==0.8.0
Django==5.2
dj-database-url==2.2.0
django-cors-headers==4.7.0
django-filter==25.1
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
filelock==3.18.0
fonttools==4.58.5
gunicorn==22.0.0
idna==3.10
Jinja2==3.1.6
lxml==5.3.2
MarkupSafe==3.0.2
packaging==25.0
pdf2image==1.17.0
pdfkit==1.0.0
pgvector==0.4.1
pillow==11.3.0
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
pytesseract==0.3.13
python-docx==1.1.2
PyYAML==6.0.2
redis==5.0.7
requests==2.32.4
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.7.0
zopfli==0.2.3.post1
celery==5.4.0
